cesizen/src/backend/user_service/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
pymongo==4.6.1
python-jose==3.3.0
requests==2.31.0